renamed userdb to userdatabase
